# docs/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ISignup": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"repeatPassword": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"repeatPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVerifyAccount": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"token": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPatchUser": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"repeatPassword": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"mobile": {
						"type": "string"
					},
					"blocked": {
						"type": "boolean"
					}
				},
				"required": [
					"email",
					"password",
					"repeatPassword",
					"token",
					"firstName",
					"lastName",
					"mobile",
					"blocked"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IForgotPassword": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILogin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGetUser": {
				"properties": {
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "questionia",
		"version": "1.0.0",
		"description": "a basic express app writen in typescript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Abdulraheem A omeiza"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/user/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"password": "123456qwerty",
											"repeat_passwork": "123456qwerty",
											"email": "peddleCustomer@peddle.com"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "email already in use"
					}
				},
				"tags": [
					"USERS"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignup"
							}
						}
					}
				}
			}
		},
		"/user/verify": {
			"post": {
				"operationId": "VerifyToken",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"201": {
						"description": "Verified successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USERS"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVerifyAccount"
							}
						}
					}
				}
			}
		},
		"/user/update-profile": {
			"post": {
				"operationId": "CreateProfile",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"201": {
						"description": "Profile updated successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USERS"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPatchUser"
							}
						}
					}
				}
			}
		},
		"/user/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"201": {
						"description": "Password reset link sent successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USERS"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IForgotPassword"
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"201": {
						"description": "Logged in successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USERS"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ILogin"
							}
						}
					}
				}
			}
		},
		"/user/user/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"201": {
						"description": "Profile fetched successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USERS"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/delete": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"201": {
						"description": "User deleted successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USERS"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGetUser"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}